Move metadata collection into a separate module (metadata_collectors)
