+++ smpl/src/analysis/mod.rs
@@ -19,6 +19,7 @@ mod type_resolver;
 mod unique_linear_cfg_traversal;
 mod analysis_context;
 mod anon_storage;
+mod metadata_collectors;
 
 pub mod error;
 pub mod metadata;
